data_pipeline/requirements.txt
pytest==8.3.3
flake8==7.1.1
discord.py==2.3.2
python-dotenv==1.0.1
supabase>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.3
pandas>=2.2.3